requirements.txt
discord.py==2.3.2
mutagen==1.47.0
python-dotenv==1.0.1
pytube==15.0.0
youtube_data_api==0.0.21
yt_dlp==2024.11.18
pynacl==1.5.0